requirements.txt
sentence-transformers
psycopg2-binary
pgvector[psycopg2]
numpy
protobuf<4
